@@ -1,5 +1,5 @@
 {
-  "name": "server-register-client",
+  "name": "spike-client",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -12,50 +12,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/cdk": "^6.4.2",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^6.4.2",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/router": "^6.0.3",
-    "@types/typescript": "^2.0.0",
-    "classlist.js": "^1.1.20150312",
-    "core-js": "^2.5.4",
-    "elastic-apm-node": "^2.16.0",
+    "@angular/animations": "~9.0.6",
+    "@angular/cdk": "^9.1.2",
+    "@angular/common": "~9.0.6",
+    "@angular/compiler": "~9.0.6",
+    "@angular/core": "~9.0.6",
+    "@angular/forms": "~9.0.6",
+    "@angular/material": "^9.1.2",
+    "@angular/platform-browser": "~9.0.6",
+    "@angular/platform-browser-dynamic": "~9.0.6",
+    "@angular/router": "~9.0.6",
+    "cookie-parser": "^1.4.5",
+    "elastic-apm-node": "^3.5.0",
     "express-serve-static-core": "^0.1.1",
-    "hammerjs": "^2.0.8",
-    "lodash": "^4.17.11",
+    "express-session": "^1.17.0",
+    "jsonwebtoken": "^8.5.1",
+    "material-design-icons": "^3.0.1",
     "ng2-avatar": "^2.3.1",
-    "nodejs": "0.0.0",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.26"
+    "ngx-avatar": "^3.7.0",
+    "passport": "^0.4.1",
+    "passport-shraga": "^1.4.0",
+    "rxjs": "~6.5.4",
+    "tsc": "^1.20150623.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.6",
-    "@angular/cli": "~7.0.4",
-    "@angular/compiler-cli": "^7.2.10",
-    "@angular/language-service": "^6.0.3",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "~0.900.6",
+    "@angular/cli": "~9.0.6",
+    "@angular/compiler-cli": "~9.0.6",
+    "@angular/language-service": "~9.0.6",
+    "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.2",
-    "ts-node": "~5.0.1",
-    "tslint": "^5.9.1",
-    "typescript": "3.1.6"
+    "karma": "~4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.3",
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "^3.7.5"
   }
 }
