@@ -1,6 +1,5 @@
-// page-not-found.component.spec
-
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
 import { PageNotFoundComponent } from './page-not-found.component';
 
 describe('PageNotFoundComponent', () => {
