@@ -2,14 +2,15 @@
 
 export const config = {
 
-//  SERVER_HOST: 'http://40.68.41.239',
-    SERVER_HOST: 'https://localhost',
-    SERVER_PORT: '3000',
-    CLIENT_PORT: '443',
-    PRIVATE_KEY_PATH: 'src/certs/files/privatekey.pem',
-    CERT_PATH: 'src/certs/files/certificate.pem',
-    ELASTIC_APM_SERVICE_NAME: 'apm_service',
-    ELASTIC_APM_SERVER_URL: 'apm_server',
-    ELASTIC_APM_SECRET_TOKEN: 'apm_secret',
-    ELASTIC_APM_ACTIVE: 'false',
+        SERVER_HOST: 'https://localhost',
+    //  SERVER_HOST: 'https://51.144.178.121',
+        SERVER_PORT: '3005',
+        CLIENT_PORT: '4200',
+        PRIVATE_KEY_PATH: 'src/certs/files/privatekey.pem',
+        CERT_PATH: 'src/certs/files/certificate.pem',
+        ELASTIC_APM_SERVICE_NAME: 'apm_service',
+        ELASTIC_APM_SERVER_URL: 'apm_server',
+        ELASTIC_APM_SECRET_TOKEN: 'apm_secret',
+        ELASTIC_APM_ACTIVE: 'false',
+        SHRAGA_URL: 'http://13.79.7.3'
 };
