@@ -1,6 +1,5 @@
-// sidenav.component.spec
-
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
 import { SidenavComponent } from './sidenav.component';
 
 describe('SidenavComponent', () => {
