@@ -1,6 +1,5 @@
-// toolbar.component.spec
-
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
 import { ToolbarComponent } from './toolbar.component';
 
 describe('ToolbarComponent', () => {
